moving fetchquizzesonrefresh to seperate file
