+++ src/utilities/Index.js
@@ -2,3 +2,5 @@ export * from './addQuiz';
 export * from './fetchQuiz';
 export * from './formatTime';
 export * from './hoursAndMinutesToseconds';
+export * from './fetchLowestResults';
+export * from './fetchQuizzesOnRefresh';
